App will now create a local database if one does not already exists and have project and part tables within it. Will automatically read tables for project information if they exist.

+++ Project Car Build Log/Support/Part.swift
@@ -14,7 +14,7 @@ class Part {
     private var cost: Float
     private var date: String // optional; default is 00/00/00
     
-    init(newNum: Int, newName: String, newCost: Float, newDate: String = "00/00/00") {
+    init(newNum: Int, newName: String, newCost: Double, newDate: String = "00/00/00") {
         self.partNum = newNum
         self.partName = newName
         self.cost = newCost
